requirements.txt
py-clob-client>=0.34.5
python-dotenv>=1.0.0
requests>=2.31.0
playwright>=1.40.0
openpyxl>=3.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0
